requirements.txt
opencv-python==4.8.0
tensorflow==2.13.0
mediapipe==0.10.0
numpy==1.24.3
streamlit==1.25.0
pillow==10.0.0
scikit-learn==1.3.0 